refactor(session-grouping): split monolithic file into a package with utils/llm/cron layers
The old api/src/agent/sessionGrouping.ts had grown to 1767 lines with
four LLM prompts, all the path-extraction code, the cron, the cron's
helpers, and the orchestration entry points all living together.
Splitting it makes responsibilities obvious and prevents the next round
of changes from drifting them back together.

Folder layout:

  api/src/agent/sessionGrouping/
    index.ts                # public surface + __testing__ barrel
    sessionGrouping.ts      # orchestration: updateSessionGroup,
                            # summariseSession, buildProjectContext
    llm/
      classifyGroup.ts
      generateSessionSummary.ts
      generateGroupDescriptionFromSummaries.ts
      index.ts
    cron/
      refreshGroupDescription.ts   # + lastRefreshedAt / refreshKey
      refreshAllSessionGroups.ts
      startGroupingCronJob.ts
      index.ts
    utils/
      stripResponseWrappers.ts
      truncateOnSentenceBoundary.ts
      formatSessionTimestamp.ts
      pathRelations.ts             # isAncestorOrEqualPath, pathsRelated
      runWithConcurrency.ts
      pathExtraction.ts            # URL stripper, local-prefix allow-list,
                                   # NON_ROOT_SEGMENTS, regex factory, etc.
      trimToProjectRoot.ts
      extractProjectPath.ts
      extractUserInputs.ts         # + stripInjectedWrappers(Rich)
      extractStoredProjectPath.ts
      findGroupByExactPath.ts
      index.ts

Largest file is now 286 lines (sessionGrouping.ts orchestration),
versus 1767 before.

Dead code removed:

  - stripInjectedWrappers (the string wrapper) — only referenced from
    __testing__; kept the rich variant the real callers use, dropped
    the wrapper. (Re-added as a thin re-export for test compatibility.)
  - fetchSiblingInputs — completely unused after the move to
    per-session summaries.
  - enrichGroupDescription — completely unused after
    generateGroupDescriptionFromSummaries replaced it. The 120 lines
    of LLM prompt + safety-net guards were carrying nothing.
  - extractAllProjectRoots — only referenced from tests and the
    __testing__ export, never from production code.
  - extractKeySubdirectories — only used inside the deleted
    enrichGroupDescription, so dead by transitivity.
  - buildDescription — deterministic fallback that classifyGroup used
    to produce; classifyGroup is now name-only and never produces a
    description.
  - Six unit tests covering those dead helpers (3 for
    extractAllProjectRoots, 3 for extractKeySubdirectories).

Public API surface is unchanged: every external import still resolves
via the package barrel (`./sessionGrouping/index.ts`). The
`__testing__` object still exposes the same helper names the tests
reach for (plus the cron helpers we already exposed).

Verified:

  - tsc build green.
  - prettier --check clean.
  - 88/88 tests pass (down from 94 because 6 dead-code tests were
    removed; all surviving tests pass unchanged).
  - One-shot cron runner against production sqlite: all 9 groups
    refreshed without issues, dominant roots and split detection work
    end-to-end exactly as before.

Net diff: 5 files changed, +1644 / -1767.
