+++ api/src/agent/sessionGrouping/llm/index.ts
@@ -1,0 +1,3 @@
+export { classifyGroup, type GroupResult } from './classifyGroup';
+export { generateSessionSummary } from './generateSessionSummary';
+export { generateGroupDescriptionFromSummaries } from './generateGroupDescriptionFromSummaries';
